# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kebelet Open API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/checkins": {
            "get": {
                "tags": [
                    "Checkin"
                ],
                "summary": "Checkin List",
                "description": "Checkin List here",
                "operationId": "checkinList",
                "parameters": [
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Paginate size",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Paginate page",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkins retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Checkin"
                ],
                "summary": "Checkin Create",
                "description": "Checkin Create here",
                "operationId": "checkinCreate",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "toilet_id"
                                ],
                                "properties": {
                                    "toilet_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checkin successfully created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/checkins/{id}": {
            "get": {
                "tags": [
                    "Checkin"
                ],
                "summary": "Checkin Retrieve",
                "description": "Checkin Retrieve here",
                "operationId": "checkinRetrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checkin id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkin retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Checkin not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Checkin"
                ],
                "summary": "Checkin Delete",
                "description": "PrCheckinoduct Delete here",
                "operationId": "checkinDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checkin id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkin deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to delete the Checkin."
                    },
                    "404": {
                        "description": "Checkin not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}